refactor(refs T30969): move DpFlyout to components root
