+++ src/components/DpFlyout/DpFlyout.stories.mdx
@@ -1,0 +1,21 @@
+import { Meta } from '@storybook/addon-docs'
+import DpFlyout from './DpFlyout'
+
+<Meta
+  title="Components/Flyout"
+  component={DpFlyout} />
+
+# Flyout
+
+The Flyout component renders a flyout that is intended to show further content.
+
+export const Template = (args, { argTypes }) => ({
+  props: Object.keys(argTypes),
+  components: { DpFlyout },
+  template: `<dp-flyout v-bind="$props" />`
+});
+
+## Roadmap
+
+- use https://popper.js.org/docs/v2/ for positioning instead of basic css - or find
+another way to make left/right offset dynamic.
