--- src/components/core/DpFlyout.stories.mdx
@@ -1,8 +1,0 @@
-# Flyout
-
-The Flyout component renders a flyout that is intended to show further content.
-
-## Roadmap
-
-- use https://popper.js.org/docs/v2/ for positioning instead of basic css - or find
-another way to make left/right offset dynamic.
